Don't pass a URL on tabs.onUpdated or tabs.onActivated
Tabs can be updated in the background without user action so we
need to get the URL of the current active tab rather than use
the URL of the tab that was updated.

(We could check to see if it is the URL currently being timed,
but it's simpler to just go ahead and clock off and on, and
it's probably not that often that a background tab is updated.)

This kind of complication may not apply to tab activation, but it
seems simpler to just make sure and get the current active tab in
that case too.  This also simplifies the code as it means the
only URLs to be passed are in blue timing mode.