Disable logging for incognito/private browsing tabs
The toolbar button is disabled and the button badge removed for
incognito/private tabs to indicate that logging is not active.

We disable these tabs on install/startup and new tab creation
to prevent visible flicker from enabled to disabled state.

Also we make sure this disabled button state is maintained by
resetting it in the clock on process, otherwise it can become
re-enabled when e.g. moving the tab to a new window.

Firefox does currently not support this via manifest.json:
https://bugzilla.mozilla.org/show_bug.cgi?id=1345474
When it does we can use that instead.